# grpcApi/server.go
package grpcApi

import (
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"lazervaultGo/configs"
	"lazervaultGo/grpcApi/middleware"
	"lazervaultGo/models"
	"lazervaultGo/pb"
	"lazervaultGo/services"
	"lazervaultGo/token"
	"lazervaultGo/worker"
	"net"
	"net/http"
	"strings"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/rs/cors"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/credentials/insecure"
	"google.golang.org/grpc/reflection"
	"google.golang.org/grpc/status"
	"gorm.io/gorm"
)

type Server struct {
	config                 *configs.Config
	db                     *gorm.DB
	tokenMaker             token.Maker
	grpcServer             *grpc.Server
	httpServer             *http.Server
	redisWorker            *worker.RedisWorker
	voiceSessionController *VoiceSessionController
	userService            services.IUserService
}

func NewServer(
	db *gorm.DB,
	config *configs.Config,
	tokenMaker token.Maker,
	redisWorker *worker.RedisWorker,
) (*Server, error) {
	if redisWorker == nil {
		return nil, errors.New("redisWorker dependency cannot be nil in NewServer")
	}

	server := &Server{
		config:      config,
		db:          db,
		tokenMaker:  tokenMaker,
		redisWorker: redisWorker,
	}

	grpcServer := grpc.NewServer(
		grpc.UnaryInterceptor(middleware.AuthInterceptor(tokenMaker)),
	)

	// Initialize services
	distributor := redisWorker.GetDistributor()
	authService := services.NewAuthService(db, config, tokenMaker, distributor)
	accountService := services.NewAccountService(db, distributor)
	recipientService := services.NewRecipientService(db)
	transferService := services.NewTransferService(db, config, distributor, recipientService, accountService)
	accountCardService := services.NewAccountCardService(db, config)
	chatService := services.NewChatService(db)
	userService := services.NewUserService(db, config, tokenMaker)
	exchangeService := services.NewExchangeService(db, distributor)
	invoiceService := services.NewInvoiceService(db)
	depositService := services.NewDepositService(db, distributor, accountService)
	withdrawalService := services.NewWithdrawalService(db, distributor)
	generateTxDataService := services.NewGenerateTxDataService(db, *config)
	generateTxDataController := NewGenerateTxDataController(*generateTxDataService, tokenMaker, db)
	txFileService := services.NewTxFileService(db)
	txFileController := NewTxFileController(txFileService, tokenMaker, db)

	// Initialize AI Chat Service
	aiChatService := services.NewAIChatService(db, config, distributor)

	// Initialize AI Chat Controller
	aiChatController := NewAIChatController(aiChatService, userService)

	// Initialize Voice Session Service
	voiceSessionService := services.NewVoiceSessionService(db, config, tokenMaker, aiChatService)

	// Initialize Voice Session Controller
	voiceSessionController := NewVoiceSessionController(voiceSessionService, userService)

	// Initialize Facial Recognition Service
	facialRecognitionService := services.NewFacialRecognitionService(config.AiServiceURL)

	// Initialize Facial Recognition Controller
	facialRecognitionController := NewFacialRecognitionController(facialRecognitionService)

	// Initialize Invoice Payment Service
	invoicePaymentService := services.NewInvoicePaymentService(db)

	// Initialize Invoice Payment Controller
	invoicePaymentController := NewInvoicePaymentController(invoicePaymentService, userService, db)

	// Initialize Tagged Invoice Service
	taggedInvoiceService := services.NewTaggedInvoiceService(db)

	// Initialize Tagged Invoice Controller
	taggedInvoiceController := NewTaggedInvoiceController(taggedInvoiceService, userService, db)

	// Initialize Invoice Notification Service
	invoiceNotificationService := services.NewInvoiceNotificationService(db)

	// Initialize Invoice Notification Controller (not implemented yet)
	_ = invoiceNotificationService

	// Initialize Insurance Service
	insuranceService := services.NewInsuranceService(db, distributor)

	// Initialize Insurance Controller
	insuranceController := NewInsuranceController(insuranceService)

	// Initialize Invoice Controller
	invoiceController := NewInvoiceController(invoiceService, userService)

	// Store controllers in server for HTTP handlers
	server.voiceSessionController = voiceSessionController
	server.userService = userService

	// Store facial recognition controller for custom HTTP handlers if needed
	_ = facialRecognitionController

	// Register gRPC services
	pb.RegisterAuthServiceServer(grpcServer, NewAuthController(authService))
	pb.RegisterUserServiceServer(grpcServer, NewUserController(server))
	pb.RegisterTransferServiceServer(grpcServer, NewTransferController(transferService, db))
	pb.RegisterAccountServiceServer(grpcServer, NewAccountController(accountService, userService))
	pb.RegisterAccountCardServiceServer(grpcServer, NewAccountCardController(accountCardService, userService))
	pb.RegisterRecipientServiceServer(grpcServer, NewRecipientController(recipientService, userService))
	pb.RegisterChatServiceServer(grpcServer, NewChatController(chatService))
	pb.RegisterExchangeServiceServer(grpcServer, NewExchangeController(exchangeService, userService))
	pb.RegisterInvoiceServiceServer(grpcServer, invoiceController)
	pb.RegisterDepositServiceServer(grpcServer, NewDepositController(depositService, userService))
	pb.RegisterWithdrawServiceServer(grpcServer, NewWithdrawalController(withdrawalService, userService))
	pb.RegisterGenerateTxDataServiceServer(grpcServer, generateTxDataController)
	pb.RegisterTxFileServiceServer(grpcServer, txFileController)
	pb.RegisterAIChatServiceServer(grpcServer, aiChatController)
	pb.RegisterVoiceSessionServiceServer(grpcServer, voiceSessionController)
	pb.RegisterFacialRecognitionServiceServer(grpcServer, facialRecognitionService)
	pb.RegisterInvoicePaymentServiceServer(grpcServer, invoicePaymentController)
	pb.RegisterTaggedInvoiceServiceServer(grpcServer, taggedInvoiceController)
	pb.RegisterInsuranceServiceServer(grpcServer, insuranceController)
	server.grpcServer = grpcServer

	// Register reflection service on gRPC server.
	reflection.Register(grpcServer)

	return server, nil
}

func (s *Server) Start() error {
	// Start gRPC server
	grpcAddr := fmt.Sprintf(":%s", s.config.GRPCServerPort)
	listener, err := net.Listen("tcp", grpcAddr)
	if err != nil {
		return fmt.Errorf("failed to listen on gRPC port %s: %w", s.config.GRPCServerPort, err)
	}
	fmt.Printf("Starting gRPC server on %s\n", grpcAddr)
	go func() {
		if err := s.grpcServer.Serve(listener); err != nil {
			panic(fmt.Sprintf("failed to serve gRPC: %v", err))
		}
	}()

	// Start HTTP gateway server
	return s.startHTTPServer()
}

func (s *Server) startHTTPServer() error {
	ctx := context.Background()
	ctx, cancel := context.WithCancel(ctx)
	defer cancel()

	gwmux := runtime.NewServeMux(
		runtime.WithIncomingHeaderMatcher(customHeaderMatcher),
		runtime.WithOutgoingHeaderMatcher(customHeaderMatcher),
	)

	// The gRPC server is now on its own port again.
	grpcDialAddr := fmt.Sprintf("localhost:%s", s.config.GRPCServerPort)
	opts := []grpc.DialOption{grpc.WithTransportCredentials(insecure.NewCredentials())}

	// Register service handlers
	if err := pb.RegisterAuthServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register auth gateway: %w", err)
	}
	if err := pb.RegisterUserServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register user gateway: %w", err)
	}
	if err := pb.RegisterTransferServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register transfer gateway: %w", err)
	}
	if err := pb.RegisterAccountServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register account gateway: %w", err)
	}
	if err := pb.RegisterAccountCardServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register account card gateway: %w", err)
	}
	if err := pb.RegisterRecipientServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register recipient gateway: %w", err)
	}
	if err := pb.RegisterChatServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register chat gateway: %w", err)
	}
	if err := pb.RegisterExchangeServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register exchange gateway: %w", err)
	}
	if err := pb.RegisterInvoiceServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register invoice gateway: %w", err)
	}
	if err := pb.RegisterDepositServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register deposit gateway: %w", err)
	}
	if err := pb.RegisterWithdrawServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register withdraw gateway: %w", err)
	}
	if err := pb.RegisterAIChatServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register ai chat gateway: %w", err)
	}

	// Register remaining service handlers
	if err := pb.RegisterGenerateTxDataServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register generate tx data gateway: %w", err)
	}
	if err := pb.RegisterTxFileServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register tx file service gateway: %w", err)
	}
	if err := pb.RegisterVoiceSessionServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register voice session service gateway: %w", err)
	}
	// Note: Invoice Payment and Tagged Invoice services don't have HTTP gateway support yet
	// To enable HTTP endpoints, add google.api.http annotations to the proto files
	if err := pb.RegisterFacialRecognitionServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register facial recognition service gateway: %w", err)
	}
	if err := pb.RegisterInsuranceServiceHandlerFromEndpoint(ctx, gwmux, grpcDialAddr, opts); err != nil {
		return fmt.Errorf("failed to register insurance gateway: %w", err)
	}

	// Create main HTTP mux for non-gRPC traffic (swagger, gateway)
	httpMux := http.NewServeMux()

	// Add Swagger handler
	httpMux.Handle("/swagger/", http.StripPrefix("/swagger/", http.FileServer(http.Dir("./swagger"))))

	// Add custom voice note handler for multipart form uploads
	httpMux.HandleFunc("/v1/voice/note/upload", func(w http.ResponseWriter, r *http.Request) {
		s.handleVoiceNoteUpload(w, r, s.voiceSessionController, s.userService)
	})

	// Add gateway handler (this should come last to catch all other routes)
	httpMux.Handle("/", gwmux)

	// Setup CORS for HTTP traffic (gateway, swagger)
	corsHandler := cors.New(cors.Options{
		AllowedOrigins: []string{"*"},
		AllowedMethods: []string{
			http.MethodGet,
			http.MethodPost,
			http.MethodPut,
			http.MethodPatch,
			http.MethodDelete,
			http.MethodOptions,
		},
		AllowedHeaders: []string{
			"Accept",
			"Content-Type",
			"Content-Length",
			"Accept-Encoding",
			"Authorization",
			"X-CSRF-Token",
			"X-Refresh-Token",
		},
		ExposedHeaders: []string{
			"Authorization",
			"X-Refresh-Token",
		},
		AllowCredentials: true,
	})

	// Wrap the HTTP mux with CORS
	corsEnabledHttpMux := corsHandler.Handler(httpMux)

	// Determine port for HTTP server
	httpPort := s.config.HTTPServerPort
	if httpPort == "" {
		httpPort = s.config.HTTPServerPort // Fallback to config for local development
	}

	// Start HTTP server
	httpAddr := fmt.Sprintf(":%s", httpPort)
	fmt.Printf("Starting HTTP gateway server on %s\n", httpAddr)

	// Store the server instance so it can be shut down gracefully if needed.
	s.httpServer = &http.Server{
		Addr:    httpAddr,
		Handler: corsEnabledHttpMux, // No longer using h2c.NewHandler
	}
	return s.httpServer.ListenAndServe()
}

func customHeaderMatcher(key string) (string, bool) {
	switch key {
	case "Authorization", "X-Refresh-Token":
		return key, true
	default:
		return runtime.DefaultHeaderMatcher(key)
	}
}

func (s *Server) Stop() {
	// Gracefully stop gRPC server
	s.grpcServer.GracefulStop()

	// Gracefully stop HTTP server
	if s.httpServer != nil {
		if err := s.httpServer.Shutdown(context.Background()); err != nil {
			fmt.Printf("Error shutting down HTTP server: %v\n", err)
		}
	}
}

func (s *Server) handleVoiceNoteUpload(w http.ResponseWriter, r *http.Request, voiceSessionController *VoiceSessionController, userService services.IUserService) {
	// Set CORS headers
	w.Header().Set("Access-Control-Allow-Origin", "*")
	w.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS")
	w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")

	// Handle preflight requests
	if r.Method == "OPTIONS" {
		w.WriteHeader(http.StatusOK)
		return
	}

	// Only allow POST method
	if r.Method != "POST" {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	// Parse multipart form (25MB max)
	err := r.ParseMultipartForm(25 << 20)
	if err != nil {
		http.Error(w, "Failed to parse multipart form", http.StatusBadRequest)
		return
	}

	// Extract JWT token from Authorization header
	authHeader := r.Header.Get("Authorization")
	if authHeader == "" {
		http.Error(w, "Authorization header is required", http.StatusUnauthorized)
		return
	}

	// Validate Bearer token format
	const bearerPrefix = "Bearer "
	if !strings.HasPrefix(authHeader, bearerPrefix) {
		http.Error(w, "Authorization header must be Bearer token", http.StatusUnauthorized)
		return
	}

	jwtToken := authHeader[len(bearerPrefix):]
	if jwtToken == "" {
		http.Error(w, "JWT token is required", http.StatusUnauthorized)
		return
	}

	// Verify and parse JWT token to get user info
	payload, err := s.tokenMaker.VerifyToken(jwtToken)
	if err != nil {
		http.Error(w, "Invalid JWT token", http.StatusUnauthorized)
		return
	}

	// Get user from token payload
	user, err := getUserByEmail(userService, s.db, payload.Email)
	if err != nil {
		http.Error(w, "User not found", http.StatusUnauthorized)
		return
	}

	userID := user.ID
	// Note: tx_history is automatically fetched by the service, not sent from frontend

	// Get audio file
	file, fileHeader, err := r.FormFile("audio_file")
	if err != nil {
		http.Error(w, "audio_file is required", http.StatusBadRequest)
		return
	}
	defer file.Close()

	// Read file content
	audioContent, err := io.ReadAll(file)
	if err != nil {
		http.Error(w, "Failed to read audio file", http.StatusInternalServerError)
		return
	}

	// Create gRPC request (tx_history will be fetched automatically by service)
	req := &pb.ProcessVoiceNoteRequest{
		TxHistory: "", // Empty - service will fetch transaction history automatically
	}

	// Create context with authentication payload (like other gRPC endpoints)
	ctx := r.Context()
	ctx = context.WithValue(ctx, middleware.AuthorizationPayloadKey, payload)
	ctx = context.WithValue(ctx, middleware.AccessTokenKey, jwtToken)

	// Call the controller multipart method with audio data
	resp, err := voiceSessionController.ProcessVoiceNoteMultipart(ctx, userID, jwtToken, audioContent, fileHeader.Filename, fileHeader.Header.Get("Content-Type"), req)
	if err != nil {
		// Handle gRPC errors
		if st, ok := status.FromError(err); ok {
			switch st.Code() {
			case codes.InvalidArgument:
				http.Error(w, st.Message(), http.StatusBadRequest)
			case codes.Unauthenticated:
				http.Error(w, st.Message(), http.StatusUnauthorized)
			case codes.Internal:
				http.Error(w, st.Message(), http.StatusInternalServerError)
			case codes.Unavailable:
				http.Error(w, st.Message(), http.StatusServiceUnavailable)
			default:
				http.Error(w, st.Message(), http.StatusInternalServerError)
			}
		} else {
			http.Error(w, "Internal server error", http.StatusInternalServerError)
		}
		return
	}

	// Set response headers
	w.Header().Set("Content-Type", "application/json")

	// Convert gRPC response to JSON
	jsonResp, err := json.Marshal(map[string]interface{}{
		"success":            resp.Success,
		"msg":                resp.Msg,
		"response":           resp.Response,
		"transcribed_text":   resp.TranscribedText,
		"processing_time_ms": resp.ProcessingTimeMs,
	})
	if err != nil {
		http.Error(w, "Failed to marshal response", http.StatusInternalServerError)
		return
	}

	// Send response
	w.WriteHeader(http.StatusOK)
	w.Write(jsonResp)
}

// Helper function to get user by ID
func getUserByID(userService services.IUserService, db *gorm.DB, userID uint) (*models.User, error) {
	// This is a simplified implementation - you may need to adjust based on your UserService interface
	user, err := models.User{}.FindById(db, userID)
	if err != nil {
		return nil, err
	}
	return user, nil
}

// Helper function to get user by email
func getUserByEmail(userService services.IUserService, db *gorm.DB, email string) (*models.User, error) {
	// Find user by email using the existing model method
	user, err := models.User{}.GetUserByEmail(db, email)
	if err != nil {
		return nil, err
	}
	return user, nil
}