# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LazerVault API",
    "description": "LazerVault Account Card API",
    "version": "1.0"
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/voice/note/upload": {
      "post": {
        "summary": "Process Voice Note",
        "description": "Upload and process an audio file via multipart form data. File should be uploaded as 'audio_file' field.",
        "operationId": "VoiceSessionService_ProcessVoiceNote",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbProcessVoiceNoteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Request to process an audio file via multipart form upload.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbProcessVoiceNoteRequest"
            }
          }
        ],
        "tags": [
          "Voice Note"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "pbProcessVoiceNoteRequest": {
      "type": "object",
      "properties": {
        "txHistory": {
          "type": "string",
          "description": "Optional JSON string containing recent transactions."
        }
      },
      "description": "Request to process an audio file via multipart form upload.",
      "title": "Process Voice Note Request"
    },
    "pbProcessVoiceNoteResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "description": "Indicates if the voice note was successfully processed."
        },
        "msg": {
          "type": "string",
          "description": "Status message related to the processing."
        },
        "response": {
          "type": "string",
          "description": "AI's response based on the voice note content."
        },
        "transcribedText": {
          "type": "string",
          "description": "Transcribed text from the audio file."
        },
        "processingTimeMs": {
          "type": "string",
          "format": "int64",
          "description": "Processing time in milliseconds."
        }
      },
      "description": "Response containing AI analysis and transcription of the voice note.",
      "title": "Process Voice Note Response"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}